refactor(server): split data_model.rs into one-file-per-extractor-pass module
data_model.rs (1568 LOC) bundled ~190 LOC of serde DTOs with seven
independent extractor passes that run as parallel rayon::join branches and
share no mutable state. Convert it into a data_model/ directory:

  types.rs        public DTOs + pub(crate) EntityJob
  mod.rs          extract_data_model orchestrator + tests + `pub use types::*`
  metadata.rs     extract_entity_metadata
  properties.rs   extract_properties / extract_property
  quantities.rs   extract_quantities / extract_quantity_value
  relationships.rs extract_relationships / extract_relationship
  associations.rs IfcRelAssociates{Classification,Material,Document} + shared helpers
  spatial.rs      build_spatial_hierarchy + recursion

Pure code motion. The eight orchestrator-called extractors gain pub(super),
EntityJob + its fields gain pub(crate); everything else stays private. The
external surface is unchanged: `pub use types::*` keeps every
crate::services::data_model::{DataModel, ...} path that parquet_data_model.rs
imports, and services/mod.rs is untouched. cargo test -p ifc-lite-server
green (21); clippy -D warnings clean.
